pyproject.toml
[tool.poetry]
name = "gc_registry"
version = "0.1.0"
description = "FEA Granular Certificate Demonstration Registry Platform"
authors = ["Connor Galbraith <connor@futureenergy.associates>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4"
anyio = "^4.4.0"
atomicwrites = "^1.4.1"
attrs = "^24.1.0"
authlib = "^1.3.1"
bcrypt = "^4.2.0"
cffi = "^1.16.0"
click = "^8.1.7"
colorama = "^0.4.6"
cryptography = "^43.0.0"
ecdsa = "^0.19.0"
exceptiongroup = "^1.0.0"
fastapi = "^0.112.0"
fastapi-utils = "^0.7.0"
greenlet = "^3.0.3"
h11 = "^0.14.0"
idna = "^3.7"
iniconfig = "^2.0.0"
itsdangerous = "^2.2.0"
jinja2 = "^3.1.4"
markupsafe = "^2.1.5"
numpy = "^2.0.1"
packaging = "^24.1"
pandas = "^2.2.2"
passlib = "^1.7.4"
pluggy = "^1.5.0"
py = "^1.11.0"
pyasn1 = "^0.6.0"
pycparser = "^2.22"
pydantic = "^2.8.2"
pyparsing = "^3.1.2"
pytest = "^8.3.2"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
pytz = "^2024.1"
pyyaml = "^6.0.1"
rsa = "^4.9"
setuptools = "^72.1.0"
six = "^1.16.0"
sniffio = "^1.3.1"
sqlalchemy = "^2.0.31"
sqlalchemy-utils = "^0.41.2"
sqlalchemy2-stubs = "^0.0.2a38"
sqlmodel = "^0.0.21"
tomli = "^2.0.1"
tqdm = "^4.66.5"
typing-extensions = "^4.12.2"
uvicorn = "^0.30.5"
wheel = "^0.44.0"
starlette = "0.37.2"
ruff = "^0.5.6"
psycopg2 = "^2.9.9"
markdown = "^3.7"
pydantic-settings = "^2.4.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
pytest-cov = "^5.0.0"
ruff = "^0.5.6"
typing-extensions = "^4.12.2"
types-sqlalchemy-utils = "^1.1.0"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240819"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
